pyproject.toml
[project]
name = "ai-gym-pt"
version = "0.1.0"
description = "A simple AI-powered personal trainer application, creating customized workout plans using LLMs with structured output."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.1",
    "langchain-openai>=1.0.2",
    "pydantic>=2.12.4",
    "python-toon>=0.1.3",
    "rich>=14.2.0",
]
